MAJOR refactors!
